extracting cards from screen file
